+++ src/controllers/AbstractAppController.js
@@ -19,7 +19,7 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-import 'jquery';
+import 'ngeo/misc/common-jquery-dependencies';
 import angular from 'angular';
 import 'angular-gettext';
 import 'angular-dynamic-locale';